+++ clients/python/src/hydra_umc_sdk/__init__.py
@@ -26,4 +26,4 @@ __all__ = [
     "evaluate_job",
 ]
 
-__version__ = "0.0.6"
+__version__ = "0.0.7"